Fix compilation warnings in plugin UI code
Cleaned up various compiler warnings to improve code quality:
- Removed unused functions (loadImageFromURL, getInitialsFromName) from Comment.cpp, Profile.cpp, and FollowersList.cpp
- Removed unused variables (buttonWidth, centerX, contentStartX, bounds)
- Fixed unused lambda parameters in NavigationStack.cpp
- Fixed implicit int-to-float conversion warnings in Auth.cpp and Upload.cpp
- Fixed variable shadowing warnings in PostCard.cpp and UserDiscovery.cpp
- Removed unnecessary include statements

All changes are non-functional and purely improve code quality.
The plugin now builds with zero warnings when the codebase is in a buildable state.

+++ plugin/src/ui/auth/Auth.cpp
@@ -3,10 +3,7 @@
 #include "../../security/InputValidation.h"
 #include "../../stores/AppStore.h"
 #include "../../util/Colors.h"
-#include "../../util/Constants.h"
-#include "../../util/Emoji.h"
 #include "../../util/Log.h"
-#include "../../util/Result.h"
 
 // ==============================================================================
 Auth::Auth(Sidechain::Stores::AppStore *store) : AppStoreComponent(store) {
@@ -393,7 +390,8 @@ void Auth::paint(juce::Graphics &g) {
     const float angleOffset = static_cast<float>(oauthAnimationFrame) * 0.15f;
 
     for (int i = 0; i < numDots; ++i) {
-      float angle = angleOffset + (i * juce::MathConstants<float>::twoPi / static_cast<float>(numDots));
+      float angle =
+          angleOffset + (static_cast<float>(i) * juce::MathConstants<float>::twoPi / static_cast<float>(numDots));
       float x = static_cast<float>(spinnerArea.getCentreX()) + std::cos(angle) * spinnerRadius;
       float y = static_cast<float>(spinnerArea.getCentreY()) + std::sin(angle) * spinnerRadius;
